requirements.txt
Flask==3.0.0
openpyxl==3.1.2
xlrd==2.0.1
xlwt==1.3.0
xlutils==2.0.0
gunicorn==21.2.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
